rename crypto.rs to encryption.rs
